fix: add logic in config change event handler
The config change event handler should manage configuration changes on its own to be able
to correctly render and apply the inferenceservice-config ConfigMap. Also on this event handler,
we are restarting the kserve-controller service, which should be restarted AFTER and only IFF the
aforementioned ConfigMap is configured correctly, that is, after the configuration for the two
deployment modes are set as expected.