feat: updated get images script without branch checkout for track/0.10 (#140)
* feat: updated get images script

Used by CVE scanning workflows and can be used to collect images URL for airgapped setup. Script is specific for track/0.10 branch to capture all images for that branch. Branch is handled in kubeflow-ci and bundle automation using bundle file.

Summary of changes:
- Re-designed get images script to retrieve all images for current branch. Branch management is done outside of this script.

NOTE: kserve-web-app is not available in 1.7 release (in track0.10). Publish job will fail: #83