fix: Parse custom images with ":" in their name
For 3 of the images (i.e. explainers), the charm attempts to separate
the image name from the tag, splitting the string in two using `:` as a
separator. However, this breaks the installation if the image name
contains that special character, which could be the case if a local
container registry address is provided for instance (e.g.
172.17.0.2:5000/kserve/alibi-explainer:latest).

Use `rsplit` to ensure that the rightmost part of the string (following
the last `:`) is parsed as the tag.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@canonical.com>