Don't use deprecated AIX Explainer image
This image has been deprecated in upstream and removed in the next
v0.11.0 version of KServe. To reduce the CVE footprint of CKF we'll also
replace this image with a dummy one for this release and urge users to
not use this image in with this version of KServe as well.

Signed-off-by: Kimonas Sotirchos <kimonas.sotirchos@canonical.com>