fix: replace kubernetes_manifests with resource_dispatcher library
The name of the library that is actually published in Charmhub is resource_dispatcher,
not kubernetes_manifests. This mismatch was causing charmcraft fetch-lib to fail as it
couldn't find the library under kubernetes_manifests.
This commit brings the correct library by fetching it with charmcraft and replaces
kubernetes_manifests where needed in the charm code.

Part of canonical/bundle-kubeflow#1209