fix: fetch kubernetes_manifests lib with charmcraft to match published lib (#304)
This repository contained a version of the kubernetes_manifestss lib that was not exactly the one that was published in
Charmhub. This caused issues when trying to fetch new versions of the library.

This commit replaces the offending file with the library that is actually published.

Part of canonical/bundle-kubeflow#1209