chore: bump controller manifests 0.13 > 0.14.1 (#300)
* chore: bump controller manifests 0.13 > 0.14.1

Updates the manifests to 0.14.1 version of kserve for CKF 1.10 release based on the changes in upstream.

The updates exclude the manifests related to LocalModel and ClusterLocalModel (for more details see release notes of v0.14.0, 0.14.1, and docs). This is due to a new operator being introduced upstream, and it will be implemented in CKF in a patch of 1.10.